+++ EmprestimoApi/Emprestimo.Business/Models/Cliente.cs
@@ -7,7 +7,7 @@
         public int Idade { get; set; }
         public string Cpf { get; set; }
         public double Score { get; set; }
-        public double SaldoAtual { get; set; }
+        public decimal SaldoAtual { get; set; }
         public Emprestimo Emprestimo { get; set; }
     }
 }
